+++ CHANGES.md
@@ -52,6 +52,9 @@ To be released.
     `PreEvaluationBlock<T>.DetermineStateRootHash(BlockChain<T>)`,
     and `PreEvaluationBlock<T>.DetermineStateRootHash(BlockChain<T>,
     out IImmutableDictionary<string, IValue>)`.  [[#3037]]
+ -  (Libplanet.Extensions.Cocona) Changed signature of
+    `ApvCommand.Verify(string?, string[]?, bool)` method to
+    `ApvCommand.Verify(string?, PublicKey[]?, bool)`.  [[#3044]]
 
 ### Backward-incompatible network protocol changes
 
@@ -73,8 +76,19 @@ To be released.
     `BlockChain<T>.EvaluateGenesis()`,
     `BlockChain<T>.DetermineBlockStateRootHash()`,
     and `BlockChain<T>.EvaluateBlock()`.  [[#3037]]
+ -  Added `PublicKey.FromHex()` static method.  [[#2709], [#3044]]
+ -  Added `PublicKey.ToHex()` method.  [[#2709], [#3044]]
 
 ### Behavioral changes
+
+ -  Added `TypeConverter` to commonly-used types.  [[#2711], [#3044]]
+
+     -  `Address` now can be converted from and to `string` through
+        `TypeConverter`.
+     -  `HashDigest<T>` now can be converted from and to `string` through
+        `TypeConverter`.
+     -  `PublicKey` now can be converted from and to `string` through
+        `TypeConverter`.
 
  -  `Transaction<T>` no more supports deserialization from JSON when it contains
     custom actions.  However, it still can be deserialized from JSON when it
@@ -117,9 +131,12 @@ To be released.
 
 ### CLI tools
 
+[#2709]: https://github.com/planetarium/libplanet/issues/2709
+[#2711]: https://github.com/planetarium/libplanet/issues/2711
 [#2986]: https://github.com/planetarium/libplanet/pull/2986
 [#2973]: https://github.com/planetarium/libplanet/issues/2973
 [#2996]: https://github.com/planetarium/libplanet/pull/2995
+[#3044]: https://github.com/planetarium/libplanet/pull/3044
 
 
 Version 0.53.2